fix some imports for the gameState refactor
